factor out into DataFrameDataset
